Checkout session retrieved from webhook event. No need for retrieve method.
Will use session info to create a reservation number for...
the user and save it in MongoDb.

In the next step; show the reservation to the user.

Create a profile page and show user details and acution...
attendance details such as reservation numbers etc in this page.

Keep going.

Remember why you started.